# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Nethesis Operation Center API
  description: REST API for Nethesis Operation Center with business hierarchy management and RBAC
  version: 1.0.0
  contact:
    name: Nethesis S.r.l.
    url: http://www.nethesis.it
    email: info@nethesis.it
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api
    description: Backend API server (port 8080)
  - url: http://localhost:8081/api
    description: Collect API server (port 8081)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from auth exchange endpoint
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication for system inventory collection
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    SearchParam:
      name: search
      in: query
      description: Search term
      required: false
      schema:
        type: string
        minLength: 1
        example: "acme"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              message:
                type: string
                example: "invalid token"
              data:
                type: object
                nullable: true
                example: null
...
    UnprocessableEntity:
      description: Unprocessable entity - business logic error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
  /accounts:
    get:
      tags:
        - Backend - Accounts
      summary: List accounts
      description: Get paginated list of user accounts with hierarchical authorization
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "accounts retrieved successfully"
                  data:
                    type: object
                    properties:
                      accounts:
                        type: array
                        items:
                          $ref: '#/components/schemas/Account'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Backend - Accounts
      summary: Create account
      description: Create a new user account with hierarchical authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "account created successfully"
                  data:
                    $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /accounts/{id}:
    get:
      tags:
        - Backend - Accounts
      summary: Get single account
      description: Get a specific account by ID with hierarchical authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            example: "usr_123456789"
      responses:
        '200':
          description: Account retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "account retrieved successfully"
                  data:
                    $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Backend - Accounts
      summary: Update account
      description: Update a user account with hierarchical authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            example: "usr_123456789"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: Account updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "account updated successfully"
                  data:
                    $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags:
        - Backend - Accounts
      summary: Delete account
      description: Delete a user account with hierarchical authorization
      parameters:
        - name: id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            example: "usr_123456789"
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "account deleted successfully"
                  data:
                    type: object
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{id}/password:
    patch:
      tags:
        - Backend - Accounts
      summary: Reset account password
      description: Reset a user account's password (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Account ID
          schema:
            type: string
            example: "usr_123456789"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "password reset successfully"
                  data:
                    type: object
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /applications:
    get:
      tags:
        - Backend - Applications
      summary: Get third-party applications
      description: Get third-party applications filtered by user's organization and user roles
      responses:
        '200':
          description: Applications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Applications retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ThirdPartyApplication'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Failed to fetch applications from Logto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/exchange:
    post:
      tags:
        - Backend - Authentication
      summary: Exchange Logto token for custom JWT
      description: Exchange a Logto access token for a custom JWT with embedded permissions
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenExchangeRequest'
      responses:
        '200':
          description: Token exchange successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "token exchange successful"
                  data:
                    $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
        - Backend - Authentication
      summary: Get current user information
      description: Get current authenticated user's profile and permissions
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "user information retrieved successfully"
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Backend - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "token refresh successful"
                  data:
                    $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /customers:
    get:
      tags:
        - Backend - Customers
      summary: List customers
      description: Get paginated list of customers (Owner + Distributor + Reseller)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "customers retrieved successfully"
                  data:
                    type: object
                    properties:
                      customers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Organization'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Backend - Customers
      summary: Create customer
      description: Create a new customer organization (Owner + Distributor + Reseller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "customer created successfully"
                  data:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /customers/{id}:
    get:
      tags:
        - Backend - Customers
      summary: Get single customer
      description: Get a specific customer by ID (Owner + Distributor + Reseller)
      parameters:
        - name: id
          in: path
          required: true
          description: Customer organization ID
          schema:
            type: string
            example: "org_456789123"
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "customer retrieved successfully"
                  data:
                    $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Backend - Customers
      summary: Update customer
      description: Update a customer organization (Owner + Distributor + Reseller)
      parameters:
        - name: id
          in: path
          required: true
          description: Customer organization ID
          schema:
            type: string
            example: "org_456789123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "customer updated successfully"
                  data:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags:
        - Backend - Customers
      summary: Delete customer
      description: Delete a customer organization (Owner + Distributor + Reseller)
      parameters:
        - name: id
          in: path
          required: true
          description: Customer organization ID
          schema:
            type: string
            example: "org_456789123"
      responses:
        '200':
          description: Customer deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "customer deleted successfully"
                  data:
                    type: object
                    nullable: true
  /distributors:
    get:
      tags:
        - Backend - Distributors
      summary: List distributors
      description: Get paginated list of distributors (Owner only)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: Distributors retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "distributors retrieved successfully"
                  data:
                    type: object
                    properties:
                      distributors:
                        type: array
                        items:
                          $ref: '#/components/schemas/Organization'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Backend - Distributors
      summary: Create distributor
      description: Create a new distributor organization (Owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
      responses:
        '201':
          description: Distributor created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "distributor created successfully"
                  data:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /distributors/{id}:
    get:
      tags:
        - Backend - Distributors
      summary: Get single distributor
      description: Get a specific distributor by ID (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          description: Distributor organization ID
          schema:
            type: string
            example: "org_123456789"
      responses:
        '200':
          description: Distributor retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "distributor retrieved successfully"
                  data:
                    $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Backend - Distributors
      summary: Update distributor
      description: Update a distributor organization (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          description: Distributor organization ID
          schema:
            type: string
            example: "org_123456789"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
      responses:
        '200':
          description: Distributor updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "distributor updated successfully"
                  data:
                    $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags:
        - Backend - Distributors
      summary: Delete distributor
      description: Delete a distributor organization (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          description: Distributor organization ID
          schema:
            type: string
            example: "org_123456789"
      responses:
        '200':
          description: Distributor deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "distributor deleted successfully"
                  data:
                    type: object
                    nullable: true
...
  /organization-roles:
    get:
      tags:
        - Backend - Roles
      summary: Get all organization roles
      description: Get all available organization roles with their IDs and descriptions
      responses:
        '200':
          description: Organization roles retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "organization roles retrieved successfully"
                  data:
                    type: object
                    properties:
                      organizationRoles:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrganizationRole'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations:
    get:
      tags:
        - Backend - Organizations
      summary: Get available organizations
      description: Get organizations the current user can assign users to, filtered by business hierarchy
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "organizations retrieved successfully"
                  data:
                    type: object
                    properties:
                      organizations:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrganizationReference'
        '401':
          $ref: '#/components/responses/Unauthorized'
...